refactor: combine all OSXKeyStateTests into one unittest
